Split screenshots file to smaller files
